Moved safe_load_tac to time_activity_curve.py in utils
